Moves validation logic out of the command
Validation logic methods were cluttering the command file.

These methods were moved to a validator in a flexible folder
structure outside the method and all validation is now performed
in a single method rather than one per attribute.
